[libc] Refactor core Linux syscalls to use syscall_wrappers
This patch initiates the refactoring of Linux syscalls as described in
the RFC (https://discourse.llvm.org/t/rfc-linux-syscall-cleanup/87248/).

It introduces a new infrastructure in `src/__support/OSUtil/linux/syscall_wrappers/`
to house header-only syscall wrappers. These wrappers utilize `ErrorOr`
to provide a consistent, type-safe interface for error handling across
the library, standardizing how syscall return values are converted into
errno-compatible Error objects.

Summary of changes:
- Created the `syscall_wrappers` directory and added `close.h`, `read.h`,
  `write.h`, and `open.h`.
- Moved the existing `getrandom.h` into the new `syscall_wrappers`
  directory and updated its callers (including HashTable/randomness.h).
- Refactored core entrypoints (`close`, `read`, `write`, `open`) to use
  the new wrappers, removing direct `syscall_impl` logic and manual
  errno setting.
- Updated `shm_open.cpp` to use the new `open` wrapper.
- Cleaned up `OSUtil/linux/fcntl.cpp` by removing redundant internal
  implementations of `open` and `close`.
- Added a developer guide in `docs/dev/syscall_wrapper_refactor.rst`
  outlining the established pattern for future migrations.
